Add GetProductByID method, refactor product handlers, and implement JSON encoding for products
