+++ product-api/handlers/delete.go
@@ -1,0 +1,38 @@
+package handlers
+
+import (
+	"my-go-project/product-api/data"
+	"net/http"
+	"strconv"
+
+	"github.com/gorilla/mux"
+)
+
+// swagger:route DELETE /products/{id} products deleteProduct
+// Deletes a product
+// responses:
+//  204: noContent
+//	501: errorResponse
+
+// DeleteProduct deletes a product from the data store
+func (p *Products) DeleteProduct(w http.ResponseWriter, r *http.Request) {
+	vars := mux.Vars(r)
+	id, err := strconv.Atoi(vars["id"])
+	p.l.Println("Handle DELETE Product", id)
+
+	if err != nil {
+		http.Error(w, "Unable to convert id", http.StatusBadRequest)
+		return
+	}
+
+	err = data.DeleteProduct(id)
+	if err == data.ErrProductNotFound {
+		http.Error(w, "Product not found", http.StatusNotFound)
+		return
+	}
+
+	if err != nil {
+		http.Error(w, "Product not found", http.StatusInternalServerError)
+		return
+	}
+}
